requirements.txt
docx2pdf==0.1.8
lxml==4.8.0
Pillow==9.0.1
python-docx==0.8.11
tqdm==4.62.3
